+++ napari_bioformats/_tests/test_pims.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-
+import sys
 import napari_bioformats
 import numpy as np
 import pytest
@@ -11,7 +11,11 @@ data = root / "sample_data"
 
 
 @pytest.mark.parametrize("fname", data.iterdir(), ids=lambda x: x.stem)
-def test_reader(fname):
+def test_reader(fname, monkeypatch):
+    # prevent usage of qtpy during tests
+    monkeypatch.setitem(sys.modules, "napari_bioformats._dialog", None)
+    monkeypatch.setitem(sys.modules, "napari_bioformats._downloader", None)
+
     reader = napari_get_reader(str(fname))
     assert callable(reader)
     ((data, meta),) = reader(fname)
